* 'master' of github.com:Jumpscale/jumpscale_core8: (35 commits)
  Check if dest exists before copying
  remove testing line
  handle connection error
  fix patch for ms1
  Correctly detect osname
  Jumpscripts python2 to python3
  Various fixes
  fix some aysfs issue, improve detection, stop aysfs on destroy
  agentcontroller8 installation works
  fix caching issue in process manager factory
  fixing jsuser list command
  jumpscripts folder is saved in syncthing with id
  change the api of the mongo cluster
  bug fix
  add telegram bot dependancy during setup
  add ays telegram bot to atyourservice
  AYS consumables can be defined
  fix _clean
  improve ssh client connection + cmd execution
  retry ssh connect
  ...

# Conflicts:
#	lib/JumpScale/tools/cuisine_/CuisineFactory.py
#	lib/JumpScale/tools/executor/ExecutorSSH.py

+++ lib/JumpScale/tools/cuisine_/Cuisine2.py
@@ -219,7 +219,8 @@ class OurCuisine():
         self.cd="/"
         self.sudomode = False
 
-        self.executor=executor
+        self.executor = executor
+        self.runid = self.id
 
         self._installer = None
         self._platformtype=None
@@ -250,10 +251,6 @@ class OurCuisine():
         self._js8sb=None
         self._dirs={}
 
-        if self.executor.type=="ssh":
-            self.runid="cuisine:%s:%s"%(self.executor.addr,self.executor.port)
-        else:
-            self.runid="cuisine:local"
         self.done=[]
 
     @property
@@ -302,12 +299,7 @@ class OurCuisine():
 
     @property
     def id(self):
-        if self._id==None:
-            if "addr" in self.executor.__dict__:
-                self._id="%s:%s"%(self.executor.addr,self.executor.port)
-            else:
-                self._id=""
-        return self._id
+        return self.executor.id
 
     @property
     def platformtype(self):
@@ -413,43 +405,39 @@ class OurCuisine():
     @property
     def dir_paths(self):
         if self._dirs == {}:
-            rc,out=self.run("js 'print(j.data.serializer.json.dumps(j.dirs.__dict__))'",showout=False,die=False,force=True,replaceArgs=False)
-            if False:#rc==0:
-                self._dirs=j.data.serializer.json.loads(out)
+            res={}
+            if 'JSBASE' in os.environ:
+                res["base"]= os.environ["JSBASE"]
             else:
-                res={}
-                if 'JSBASE' in os.environ:
-                    res["base"]= os.environ["JSBASE"]
-                else:
-                    if self.isMac:
-                        res["base"]= "%s/opt/jumpscale8/"%os.environ["HOME"]
-                    else:
-                        res["base"]= "/opt/jumpscale8/"
                 if self.isMac:
-                    res["codeDir"]= "%s/opt/code/"%os.environ["HOME"]
+                    res["base"]= "%s/opt/jumpscale8/"%os.environ["HOME"]
                 else:
-                    res["codeDir"]= "/opt/code/"
-                if self.isMac:
-                    res["varDir"]= "%s/optvar/"%os.environ["HOME"]
-                else:
-                    res["varDir"]= "/optvar/"
-                res["appDir"]="%s/apps"%res["base"]
-                res["binDir"]="%s/bin"%res["base"]
-                res["cfgDir"]="%s/cfg"%res["varDir"]
-                res["jsLibDir"]="%s/lib/JumpScale/"%res["base"]
-                res["libDir"]="%s/lib/"%res["base"]
-                res["homeDir"]=os.environ["HOME"]
-                res["logDir"]="%s/log"%res["varDir"]
-                res["pidDir"]="%s/pid"%res["varDir"]
-                res["tmpDir"]="%s/tmp"%res["varDir"]
-                self._dirs=res
+                    res["base"]= "/opt/jumpscale8/"
+            if self.isMac:
+                res["codeDir"]= "%s/opt/code/"%os.environ["HOME"]
+            else:
+                res["codeDir"]= "/opt/code/"
+            if self.isMac:
+                res["varDir"]= "%s/optvar/"%os.environ["HOME"]
+            else:
+                res["varDir"]= "/optvar/"
+            res["appDir"]="%s/apps"%res["base"]
+            res["binDir"]="%s/bin"%res["base"]
+            res["cfgDir"]="%s/cfg"%res["varDir"]
+            res["jsLibDir"]="%s/lib/JumpScale/"%res["base"]
+            res["libDir"]="%s/lib/"%res["base"]
+            res["homeDir"]=os.environ["HOME"]
+            res["logDir"]="%s/log"%res["varDir"]
+            res["pidDir"]="%s/pid"%res["varDir"]
+            res["tmpDir"]="%s/tmp"%res["varDir"]
+            self._dirs=res
 
         if self.isMac:
             self._dirs["optDir"]= "%s/opt/"%os.environ["HOME"]
         else:
             self._dirs["optDir"]= "/opt/"
 
-        self._dirs["goDir"]= "%sgo/"%self._dirs["optDir"]
+        self._dirs["goDir"]= "%sgo/"%self._dirs["varDir"]
 
         return self._dirs
 
@@ -653,7 +641,6 @@ class OurCuisine():
             assert self.file_exists(location), "cuisine.file_read: file does not exists {0}".format(location)
         elif not self.file_exists(location):
             return default
-
 
         frame = self.file_base64(location)
 
@@ -914,11 +901,12 @@ class OurCuisine():
     def file_copy(self, source, dest, recursive=False):
         source=self.cuisine.args_replace(source)
         dest=self.cuisine.args_replace(dest)
-        cmd = "cp -v "
-        if recursive:
-            cmd += "-r "
-        cmd += '%s %s' % (source, dest)
-        self.run(cmd)
+        if not self.file_exists(dest):
+            cmd = "cp -v "
+            if recursive:
+                cmd += "-r "
+            cmd += '%s %s' % (source, dest)
+            self.run(cmd)
 
     @actionrun(action=False,force=False)
     def file_move(self, source, dest, recursive=False):
@@ -935,11 +923,13 @@ class OurCuisine():
     def file_base64(self,location):
         """Returns the base64-encoded content of the file at the given location."""
         location=self.cuisine.args_replace(location)
+        sudomode = self.sudomode
         res=self.run("cat {0} | python3 -c 'import sys,base64;sys.stdout.write(base64.b64encode(sys.stdin.read().encode()).decode())'".format(shell_safe((location))),debug=False,checkok=False,showout=False)
         if res.find("command not found")!=-1:
             #print could not find python need to install
             self.cuisine.package.install("python3.5")
             res=self.run("cat {0} | python3 -c 'import sys,base64;sys.stdout.write(base64.b64encode(sys.stdin.read().encode()).decode())'".format(shell_safe((location))),debug=False,checkok=False,showout=False)
+        self.sudomode = sudomode
         return res
 
         # else:
@@ -1109,13 +1099,24 @@ class OurCuisine():
     # CORE
     # -----------------------------------------------------------------------------
 
+    def _clean(self, output):
+        if self.sudomode and hasattr(self.executor, 'login'):
+            dirt = '[sudo] password for %s: ' % self.executor.login
+            if output.find(dirt) != -1:
+                output = output.lstrip(dirt)
+        return output
+
+    @actionrun()
+    def set_sudomode(self):
+        self.sudomode = True
+
     @actionrun()
     def sudo(self, cmd, die=True,showout=True):
         if not self.isMac:
-            cmd=self.cuisine.args_replace(cmd)
-            passwd = self.executor.passwd if hasattr(self.executor, "passwd") else ''
-            cmd2 = 'echo %s | sudo -S bash -c "%s"' % (passwd, cmd.replace("$", "\$"))
-            return self.run(cmd2, die=die,showout=showout)
+            sudomode = self.sudomode
+            self.sudomode = True
+            return self.run(cmd, die=die,showout=showout)
+            self.sudomode = sudomode
         else:
             return self.run(cmd, die=die,showout=showout)
 
@@ -1145,8 +1146,9 @@ class OurCuisine():
             cmd = 'echo %s | sudo -S bash -c "%s"' % (passwd, cmd)
 
         rc,out=self.executor.execute(cmd,checkok=checkok, die=False, combinestdr=True,showout=showout)
+        out = self._clean(out)
 
-        if rc>0:
+        if rc:
             items2check=["sudo","wget","curl","git","openssl"]
             next=True
             while next==True:
@@ -1234,6 +1236,7 @@ class OurCuisine():
             self.file_write(location=path, content=content, mode=0o770)
 
         rc,out=self.run("bash %s"%path,showout=True,die=False)
+        out = self._clean(out)
 
         self.file_unlink(path)
